+++ changelog.md
@@ -1,2 +1,6 @@
 # 1.0.0
 - Initial release
+
+# 1.1.0
+- Add animations to ball, robot and spider
+- Ship, UFO and jetpack now have the cube inside;